# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMLAB API",
        "description": "Dokumentasi REST API untuk Sistem Informasi Manajemen Laboratorium (SIMLAB) - Institut Teknologi Kalimantan",
        "version": "1.0.0"
    },
    "paths": {
        "/api/kepala-lab/ruangan": {
            "post": {
                "tags": [
                    "Admin Master Data (Kepala Lab)"
                ],
                "summary": "Tambah ruangan baru",
                "operationId": "863d69e063e45904d1d2b8efd66ce77e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_ruangan",
                                    "kapasitas"
                                ],
                                "properties": {
                                    "nama_ruangan": {
                                        "type": "string",
                                        "example": "Lab. Fisika Dasar A"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Laboratorium untuk praktikum fisika dasar",
                                        "nullable": true
                                    },
                                    "kapasitas": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "status_aktif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ruangan berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ruangan berhasil ditambahkan."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama_ruangan": {
                                                    "type": "string",
                                                    "example": "Lab. Fisika Dasar A"
                                                },
                                                "kapasitas": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "status_aktif": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/kepala-lab/ruangan/{id}": {
            "put": {
                "tags": [
                    "Admin Master Data (Kepala Lab)"
                ],
                "summary": "Update data ruangan",
                "operationId": "813e72a34241fe2a6827b2c335fa9a4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_ruangan": {
                                        "type": "string",
                                        "example": "Lab. Fisika Dasar B"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Update deskripsi",
                                        "nullable": true
                                    },
                                    "kapasitas": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "status_aktif": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ruangan berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ruangan berhasil diperbarui."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama_ruangan": {
                                                    "type": "string",
                                                    "example": "Lab. Fisika Dasar B"
                                                },
                                                "kapasitas": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Master Data (Kepala Lab)"
                ],
                "summary": "Hapus ruangan",
                "description": "Ruangan yang sudah memiliki riwayat pengajuan tidak bisa dihapus, hanya bisa dinonaktifkan (`status_aktif: false`).",
                "operationId": "b25361cebdce5700047a8903d59338b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ruangan berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ruangan berhasil dihapus."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Ruangan tidak bisa dihapus karena memiliki riwayat pengajuan"
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/kepala-lab/alat": {
            "post": {
                "tags": [
                    "Admin Master Data (Kepala Lab)"
                ],
                "summary": "Tambah alat baru",
                "operationId": "54ff579170fc17569e12be927483e055",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_alat",
                                    "satuan",
                                    "total_stok"
                                ],
                                "properties": {
                                    "nama_alat": {
                                        "type": "string",
                                        "example": "Osiloskop Digital"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Alat ukur sinyal listrik",
                                        "nullable": true
                                    },
                                    "satuan": {
                                        "type": "string",
                                        "example": "unit"
                                    },
                                    "total_stok": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "status_aktif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alat berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alat berhasil ditambahkan."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama_alat": {
                                                    "type": "string",
                                                    "example": "Osiloskop Digital"
                                                },
                                                "satuan": {
                                                    "type": "string",
                                                    "example": "unit"
                                                },
                                                "total_stok": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
...
}